// src/types/enums.ts
export enum UserRole {
  USER = 'user',
  STUDENT = 'student',
  TEACHER = 'teacher',
  ADMIN = 'admin',
}

export enum GenderEnum {
  MALE = 'male',
  FEMALE = 'female',
}

export enum QuestionTypes {
  CHOICE = 'Multiple choice',
  TRUE_FALSE = 'True/false',
  SHORT = 'Short answer',
  COMPLET = 'Completion',
  ESSAYS = 'Essays',
}